python-backend/requirements.txt
# python-backend/requirements.txt
Flask
Pillow
transformers
torch
flask_cors